+++ rust/main/src/ast/c14n.rs
@@ -129,6 +129,9 @@ pub fn c14n_serialise_ast<T: Write>(out: &mut T, node: &NodeData) -> std::io::Re
             out.write_all(code)?;
             out.write_all(b"?>")?;
         }
+        NodeData::RcdataContent { content } => {
+            out.write_all(content)?;
+        }
         NodeData::ScriptOrStyleContent { code, .. } => {
             out.write_all(code)?;
         }