+++ rust/main/src/tag/mod.rs
@@ -1,0 +1,12 @@
+use aho_corasick::AhoCorasick;
+use aho_corasick::AhoCorasickBuilder;
+use lazy_static::lazy_static;
+
+lazy_static! {
+    pub static ref TAG_TEXTAREA_END: AhoCorasick = AhoCorasickBuilder::new()
+        .ascii_case_insensitive(true)
+        .build(&["</textarea"]);
+    pub static ref TAG_TITLE_END: AhoCorasick = AhoCorasickBuilder::new()
+        .ascii_case_insensitive(true)
+        .build(&["</title"]);
+}